TAG225 2020/06/01 2nd   Core     ・MML Analyzer:Yコマンドを使用するとキャストに失敗するバグを修正。     (またしてもさしゅさんからご指摘いただきました。ありがとうございます！(T_T)

+++ mml2vgm/Core/MMLAnalyze.cs
@@ -1294,8 +1294,8 @@ namespace Core
         private void CmdY(partWork pw,MML mml)
         {
             int n = -1;
-            byte dat = 0;
-            byte op = 0;
+            int dat = 0;
+            int op = 0;
             pw.incPos();
 
             char c = pw.getChar();
